+++ app/controllers/editor/projects_controller.rb
@@ -9,6 +9,10 @@ class Editor::ProjectsController < ApplicationController
   def settings
   end
 
+  # GET /editor/projects/1/advanced
+  def advanced
+  end
+
   # GET /editor/projects/1/invite
   def invite
   end